pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gneiss-engine"
version = "0.2.0"
description = "A rock-solid, high-performance Python core for metamorphic image manipulation"
readme = "README.md"
requires-python = ">=3.8,<3.13"
license = {text = "MIT"}
authors = [
    {name = "Gneiss-Engine Contributors", email = "your.email@example.com"}
]
maintainers = [
    {name = "Gneiss-Engine Maintainers", email = "maintainers@example.com"}
]
keywords = ["image-processing", "batch-processing", "image-manipulation", "pillow", "computer-vision"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",

    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Graphics",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Operating System :: OS Independent",
]
dependencies = [
    "Pillow>=9.0.0,<10.0.0",
    "tqdm>=4.62.0",
    "psutil>=5.9.0",
    "Pillow-SIMD>=9.0.0,<10.0.0; platform_machine == 'x86_64'",
]

[project.urls]
"Homepage" = "https://github.com/yourusername/Gneiss-Engine"
"Bug Tracker" = "https://github.com/yourusername/Gneiss-Engine/issues"

[project.optional-dependencies]
dev = [
    "pytest>=6.0.0",
    "pytest-cov>=2.12.0",
    "black>=21.5b2",
    "isort>=5.9.0",
    "flake8>=3.9.0",
    "mypy>=0.910",
    "types-Pillow",
]
avif = ["pillow-avif>=1.0.0"]
heic = ["pillow-heif>=0.10.0"]
all = ["pillow-avif>=1.0.0", "pillow-heif>=0.10.0"]

[tool.setuptools]
packages = {find = {where = ["."]}}
include-package-data = true

[tool.setuptools.package-dir]
gneiss = "gneiss"

[tool.black]
line-length = 88
target-version = ['py38']
include = "\\.pyi?$"
exclude = '''
/(\.
|_build
|buck-out
|build
|dist
|\.git
|\.hg
|\.mypy_cache
|\.nox
|\.tox
|\.venv
|venv
|\.svn
|CVS
|__pycache__
|\.pytest_cache
|examples/sample_images
)/
'''

[tool.isort]
profile = "black"
line_length = 88
skip_gitignore = true

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "W503"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "--cov=gneiss --cov-report=term-missing --cov-fail-under=90"

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true